Update enemyActions script
Add wait to patrolling enemy, have a public int to determine how many
frames the enemy will wait

Add chase enemy, which starts chase at some set distance and chases
until robber is too far away.  Add attack function if robber is between
2 and 6 units away for chasers.  (Should expand to patrol units too)

Add a bool to determine what type of attack is used.  If enemy is
mobster class, it will attempt to damage the robber.  if not, the enemy
will attempt to tackle the robber.